Activate and de-activate update editor on the objective page
